refactor: fix inconstant 2D prefix
