+++ src/engine/script/physics2d/ObjectAttacher2D.ts
@@ -2,7 +2,7 @@ import { Component } from "../../hierarchy_object/Component";
 import { GameObject } from "../../hierarchy_object/GameObject";
 
 /** @internal */
-export class Object2DAttacher extends Component {
+export class ObjectAttacher2D extends Component {
     private _target: GameObject|null = null;
 
     public set target(value: GameObject) {
