+++ src/engine/script/physics2d/collider/CircleCollider2D.ts
@@ -4,7 +4,7 @@ import type { Shape } from "../../../../box2d.ts/build/index";
 import { CircleShape } from "../../../../box2d.ts/build/index";
 import type { GameObject } from "../../../hierarchy_object/GameObject";
 import { CssHtmlElementRenderer } from "../../render/CssHtmlElementRenderer";
-import { Object2DAttacher } from "../Object2DAttacher";
+import { ObjectAttacher2D } from "../ObjectAttacher2D";
 import { getOrCreatePhysicsDebugRenderObject } from "../PhysicsDebugRender";
 import { Collider2D } from "./Collider2D";
 
@@ -18,8 +18,8 @@ export class CircleCollider2D extends Collider2D {
 
     private updateDebugDraw(): void {
         if (this._debugDraw) {
-            let objectAttacher = this.gameObject.getComponent(Object2DAttacher);
-            if (!objectAttacher) objectAttacher = this.gameObject.addComponent(Object2DAttacher)!;
+            let objectAttacher = this.gameObject.getComponent(ObjectAttacher2D);
+            if (!objectAttacher) objectAttacher = this.gameObject.addComponent(ObjectAttacher2D)!;
 
             if (this._debugObject) {
                 const renderer = this._debugObject.getComponentInChildren(CssHtmlElementRenderer)!;
