+++ src/engine/script/physics2d/collider/EdgeCollider2D.ts
@@ -6,8 +6,8 @@ import type { GameObject } from "../../../hierarchy_object/GameObject";
 import { PrefabRef } from "../../../hierarchy_object/PrefabRef";
 import type { ReadonlyVector2 } from "../../../math/ReadonlyVector2";
 import { Color } from "../../../render/Color";
-import { Css2DEdgeRenderer } from "../../render/Css2DEdgeRenderer";
-import { Object2DAttacher } from "../Object2DAttacher";
+import { CssEdgeRenderer } from "../../render/CssEdgeRenderer";
+import { ObjectAttacher2D } from "../ObjectAttacher2D";
 import { getOrCreatePhysicsDebugRenderObject } from "../PhysicsDebugRender";
 import { Collider2D } from "./Collider2D";
 
@@ -22,29 +22,29 @@ export class EdgeCollider2D extends Collider2D {
     private _edgeRadius = 0;
     private _debugDraw = true;
     private _debugObject: GameObject|null = null;
-    private _debugRenderer: Css2DEdgeRenderer|null = null;
+    private _debugRenderer: CssEdgeRenderer|null = null;
 
     private updateDebugDraw(): void {
         if (this._debugDraw) {
-            let objectAttacher = this.gameObject.getComponent(Object2DAttacher);
-            if (!objectAttacher) objectAttacher = this.gameObject.addComponent(Object2DAttacher)!;
+            let objectAttacher = this.gameObject.getComponent(ObjectAttacher2D);
+            if (!objectAttacher) objectAttacher = this.gameObject.addComponent(ObjectAttacher2D)!;
             
             if (this._debugObject) {
                 this._debugRenderer!.points = this._points;
             } else {
                 const physicsDebugRenderObject = getOrCreatePhysicsDebugRenderObject(this.engine);
-                const debugRenderer = new PrefabRef<Css2DEdgeRenderer>();
+                const debugRenderer = new PrefabRef<CssEdgeRenderer>();
                 this._debugObject = physicsDebugRenderObject.addChildFromBuilder(
                     this.engine.instantiater.buildGameObject(this.gameObject.name + "_debug_edge")
                         .withChild(this.engine.instantiater.buildGameObject("debug_edge", new Vector3(this.offset.x, this.offset.y, 200))
-                            .withComponent(Css2DEdgeRenderer, c => {
+                            .withComponent(CssEdgeRenderer, c => {
                                 c.points = this._points;
                                 c.viewScale = 0.01;
                                 c.edgeWidth = 2;
                                 c.edgeColor = new Color(1, 1, 0, 0.3);
                                 c.pointerEvents = false;
                             })
-                            .getComponent(Css2DEdgeRenderer, debugRenderer)));
+                            .getComponent(CssEdgeRenderer, debugRenderer)));
                 
                 this._debugRenderer = debugRenderer.ref;
                 objectAttacher!.target = this._debugObject;
