+++ src/engine/script/render/CssPolygonRenderer2D.ts
@@ -8,12 +8,12 @@ import { ReadonlyColor } from "../../render/ReadonlyColor";
 import { CssRenderer } from "./CssRenderer";
 
 /**
- * css 2d edge renderer
+ * css2d polygon renderer
  * 
- * this renderer use svg tag to render edge
+ * this renderer use svg tag to render polygon
  */
-export class Css2DEdgeRenderer extends CssRenderer<HTMLDivElement> {
-    private _svgElement: SVGPolylineElement|null = null;
+export class CssPolygonRenderer2D extends CssRenderer<HTMLDivElement> {
+    private _svgElement: SVGPolygonElement|null = null;
     private _points: Vector2[] = [
         new Vector2(-2, -2),
         new Vector2(2, -2),
@@ -22,9 +22,9 @@ export class Css2DEdgeRenderer extends CssRenderer<HTMLDivElement> {
     ];
     private _width = 4;
     private _height = 4;
-    private readonly _edgeColor = Color.fromHex("#00FF00");
-    private _edgeWidth = 0;
-    private static readonly _safebound = 0.1;
+    private readonly _color = Color.fromHex("#39C5BB");
+    private readonly _borderColor = Color.fromHex("#00FF00");
+    private _borderWidth = 0;
 
     protected override renderInitialize(): void {
         if (!this.htmlElement) {
@@ -33,14 +33,29 @@ export class Css2DEdgeRenderer extends CssRenderer<HTMLDivElement> {
             svgElement.setAttribute("width", "100%");
             svgElement.setAttribute("height", "100%");
             
-            const svgPolyLine = document.createElementNS("http://www.w3.org/2000/svg", "polyline");
-            svgPolyLine.style.fill = "none";
-            svgPolyLine.style.stroke = this._edgeColor.toHexWithAlpha();
-            svgPolyLine.style.strokeWidth = this._edgeWidth + "px";
-            
-            svgElement.appendChild(svgPolyLine);
+            //blur test code
+            // const defs = document.createElementNS("http://www.w3.org/2000/svg", "defs");
+            // const filter = document.createElementNS("http://www.w3.org/2000/svg", "filter");
+            // filter.setAttribute("id", "blur");
+            // filter.setAttribute("x", "-50%");
+            // filter.setAttribute("y", "-50%");
+            // filter.setAttribute("width", "200%");
+            // filter.setAttribute("height", "200%");
+            // const feGaussianBlur = document.createElementNS("http://www.w3.org/2000/svg", "feGaussianBlur");
+            // feGaussianBlur.setAttribute("in", "SourceGraphic");
+            // feGaussianBlur.setAttribute("stdDeviation", "100");
+            // filter.appendChild(feGaussianBlur);
+            // defs.appendChild(filter);
+            // svgElement.appendChild(defs);
+
+            const svgPolygon = document.createElementNS("http://www.w3.org/2000/svg", "polygon");
+            svgPolygon.style.fill = this._color.toHexWithAlpha();
+            svgPolygon.style.stroke = this._borderColor.toHexWithAlpha();
+            svgPolygon.style.strokeWidth = this._borderWidth + "px";
+            // svgPolygon.style.filter = "url(#blur)";
+            svgElement.appendChild(svgPolygon);
             this.htmlElement.appendChild(svgElement);
-            this._svgElement = svgPolyLine;
+            this._svgElement = svgPolygon;
             const css3DObject = this.initializeBaseComponents(false);
             
             Transform.updateRawObject3DWorldMatrixRecursively(css3DObject);
@@ -85,8 +100,8 @@ export class Css2DEdgeRenderer extends CssRenderer<HTMLDivElement> {
             if (Math.abs(point.x) > maxX) maxX = Math.abs(point.x);
             if (Math.abs(point.y) > maxY) maxY = Math.abs(point.y);
         }
-        this._width = maxX * 2 + this._edgeWidth * 0.03 + Css2DEdgeRenderer._safebound * 2;
-        this._height = maxY * 2 + this._edgeWidth * 0.03 + Css2DEdgeRenderer._safebound * 2;
+        this._width = maxX * 2 + this._borderWidth * 0.03;
+        this._height = maxY * 2 + this._borderWidth * 0.03;
         this.htmlElement!.style.width = (this._width / this.viewScale) + "px";
         this.htmlElement!.style.height = (this._height / this.viewScale) + "px";
     }
@@ -109,14 +124,14 @@ export class Css2DEdgeRenderer extends CssRenderer<HTMLDivElement> {
     }
 
     /**
-     * edge points (default: [(-2, -2), (2, -2), (2, 2), (-2, 2)])
+     * polygon points (default: [(-2, -2), (2, -2), (2, 2), (-2, 2)])
      */
     public get points(): readonly ReadonlyVector2[] {
         return this._points;
     }
 
     /**
-     * edge points (default: [(-2, -2), (2, -2), (2, 2), (-2, 2)])
+     * polygon points (default: [(-2, -2), (2, -2), (2, 2), (-2, 2)])
      */
     public set points(value: readonly ReadonlyVector2[]) {
         this._points.length = 0;
@@ -132,7 +147,7 @@ export class Css2DEdgeRenderer extends CssRenderer<HTMLDivElement> {
     }
 
     /**
-     * set edge points to regular polygon
+     * set polygon points to regular polygon
      * @param sides number of sides
      * @param radius radius of polygon
      */
@@ -149,34 +164,51 @@ export class Css2DEdgeRenderer extends CssRenderer<HTMLDivElement> {
     }
 
     /**
-     * edge color (default: "#00FF00")
+     * color (default: "#39C5BB")
      */
-    public get edgeColor(): ReadonlyColor {
-        return this._edgeColor;
+    public get color(): ReadonlyColor {
+        return this._color;
     }
 
     /**
-     * edge color (default: "#00FF00")
+     * color (default: "#39C5BB")
      */
-    public set edgeColor(value: ReadonlyColor) {
-        this._edgeColor.copy(value);
+    public set color(value: ReadonlyColor) {
+        this._color.copy(value);
+        if (this.htmlElement) {
+            this._svgElement!.style.fill = value.toHexWithAlpha();
+        }
+    }
+
+    /**
+     * border color (default: "#00FF00")
+     */
+    public get borderColor(): ReadonlyColor {
+        return this._borderColor;
+    }
+
+    /**
+     * border color (default: "#00FF00")
+     */
+    public set borderColor(value: ReadonlyColor) {
+        this._borderColor.copy(value);
         if (this.htmlElement) {
             this._svgElement!.style.stroke = value.toHexWithAlpha();
         }
     }
 
     /**
-     * edge width (default: 0)
+     * border width (default: 0)
      */
-    public get edgeWidth(): number {
-        return this._edgeWidth;
+    public get borderWidth(): number {
+        return this._borderWidth;
     }
 
     /**
-     * edge width (default: 0)
+     * border width (default: 0)
      */
-    public set edgeWidth(value: number) {
-        this._edgeWidth = value;
+    public set borderWidth(value: number) {
+        this._borderWidth = value;
         if (this.htmlElement) {
             this._svgElement!.style.strokeWidth = value + "px";
         }
