+++ test/index.test.ts
@@ -95,9 +95,9 @@ describe("index Test", () => {
     it("CameraType export", () => expect(Index.CameraType).toBeDefined());
     it("Camera export", () => expect(Index.Camera).toBeDefined());
     it("CameraRelativeZaxisSorter export", () => expect(Index.CameraRelativeZaxisSorter).toBeDefined());
-    it("Css2DLineRenderer export", () => expect(Index.Css2DLineRenderer).toBeDefined());
-    it("Css2DEdgeRenderer export", () => expect(Index.Css2DEdgeRenderer).toBeDefined());
-    it("Css2DPolygonRenderer export", () => expect(Index.Css2DPolygonRenderer).toBeDefined());
+    it("CssLineRenderer export", () => expect(Index.CssLineRenderer).toBeDefined());
+    it("CssEdgeRenderer2D export", () => expect(Index.CssEdgeRenderer).toBeDefined());
+    it("CssPolygonRenderer2D export", () => expect(Index.CssPolygonRenderer2D).toBeDefined());
     it("CssHtmlElementRenderer export", () => expect(Index.CssHtmlElementRenderer).toBeDefined());
     it("CssIframeRenderer export", () => expect(Index.CssIframeRenderer).toBeDefined());
     it("CssRenderer export", () => expect(Index.CssRenderer).toBeDefined());
